Move parseDictHeader to tools/http.js
